Extract Date support from stencil and use Date extension
This does not include any packaging mechanism for the Date extension.
The code needs to be embedded in the client manually.
