+++ Sources/AutoJSONSerialization/Extensions/AutoJSONSerializable.generated.swift
@@ -5,49 +5,6 @@
 
 import Foundation
 
-enum JSONDateFormatter {
-    static func date(from string: String) -> Date? {
-        if #available(iOS 10.0, macOS 10.12, *) {
-            // [HACK] workaround for unsupported milliseconds part in ISO8601DateFormatter.
-            // "1985-04-12T23:20:50Z" is supported where "1985-04-12T23:20:50.678Z" is not.
-            // So this removes the ".678" part.
-            let dateString = string.replacingOccurrences(of: "\\.\\d+", with: "", options: .regularExpression)
-            // [/HACK]
-            return isoDateFormatter.date(from: dateString)
-        } else {
-            return dateFormatter.date(from: string)
-        }
-    }
-
-    static func string(from date: Date) -> String {
-        if #available(iOS 10.0, macOS 10.12, *) {
-            return isoDateFormatter.string(from: date)
-        } else {
-            return dateFormatter.string(from: date)
-        }
-    }
-
-    @available(iOS 10.0, macOS 10.12, *)
-    private static let isoDateFormatter: ISO8601DateFormatter = {
-        let formatter = ISO8601DateFormatter()
-        formatter.formatOptions = .withInternetDateTime
-        return formatter
-    }()
-
-    private static let dateFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "en_US_POSIX")
-        formatter.dateFormat = "yyyy-MM-dd'T'HH:mm:ssZZZZZ"
-        return formatter
-    }()
-}
-
-extension Date {
-    func iso8601String() -> String {
-        return JSONDateFormatter.string(from: self)
-    }
-}
-
 // MARK: - AutoJSONSerializable for classes, protocols, structs
 
 // MARK: - ArrayProperty AutoJSONSerializable
@@ -78,7 +35,7 @@ extension BasicTypesArrayProperty: JSONSerializable {
 extension DateArrayProperty: JSONSerializable {
     internal func toJSONObject() -> Any {
         var jsonObject = [String: Any]()
-        let dateArray = self.dateArray.map { $0.iso8601String() }
+        let dateArray = self.dateArray.map { $0.toJSONObject() }
         jsonObject["dateArray"] = dateArray
         return jsonObject
     }
@@ -88,9 +45,9 @@ extension DateArrayProperty: JSONSerializable {
 extension DateProperty: JSONSerializable {
     internal func toJSONObject() -> Any {
         var jsonObject = [String: Any]()
-        let date = self.date.iso8601String()
+        let date = self.date.toJSONObject()
         jsonObject["date"] = date
-        let optionalDate = self.optionalDate?.iso8601String()
+        let optionalDate = self.optionalDate?.toJSONObject()
         jsonObject["optional_date"] = optionalDate
         return jsonObject
     }
@@ -224,7 +181,7 @@ extension StringEnumProperty: JSONSerializable {
 extension TypealiasedDateArrayProperty: JSONSerializable {
     internal func toJSONObject() -> Any {
         var jsonObject = [String: Any]()
-        let momentArray = self.momentArray.map { $0.iso8601String() }
+        let momentArray = self.momentArray.map { $0.toJSONObject() }
         jsonObject["momentArray"] = momentArray
         return jsonObject
     }
@@ -234,9 +191,9 @@ extension TypealiasedDateArrayProperty: JSONSerializable {
 extension TypealiasedDateProperty: JSONSerializable {
     internal func toJSONObject() -> Any {
         var jsonObject = [String: Any]()
-        let momentInTime = self.momentInTime.iso8601String()
+        let momentInTime = self.momentInTime.toJSONObject()
         jsonObject["momentInTime"] = momentInTime
-        let optionalMomentInTime = self.optionalMomentInTime?.iso8601String()
+        let optionalMomentInTime = self.optionalMomentInTime?.toJSONObject()
         jsonObject["optionalMomentInTime"] = optionalMomentInTime
         return jsonObject
     }
