+++ Templates/AutoJSONDeserializable.stencil
@@ -13,8 +13,8 @@ extension {{ type.name }}: JSONDeserializable {
         {% macro Optional arg %}guard {{ arg }} else { return nil }{% endmacro %}
         {% for variable in type.storedVariables %}
         {% ifnot variable.isArray %}
-        {% set castType %}{% if variable.type.implements.AutoJSONDeserializable or variable.type.implements.JSONDeserializable %}{% else %}{% if variable.actualTypeName.unwrappedTypeName == "Date" %}String{% else %}{% if variable.type.kind == "enum" and variable.type.rawTypeName %}{{ variable.type.rawTypeName.name }}{% else %}{{ variable.unwrappedTypeName }}{% endif %}{% endif %}{% endif %}{% endset %}
-        {% set itemOperation %}{% if variable.type.implements.AutoJSONDeserializable or variable.type.implements.JSONDeserializable %}.flatMap({{ variable.unwrappedTypeName }}.init(JSONObject:)){%else%}{% if variable.actualTypeName.unwrappedTypeName == "Date" %}.flatMap(JSONDateFormatter.date(from:)){% else %}{% if variable.type.kind == "enum" and variable.type.rawTypeName %}.flatMap({ {{ variable.actualTypeName.unwrappedTypeName }}(rawValue: $0) }){% endif %}{% endif %}{% endif %}{% endset %}
+        {% set castType %}{% if variable.type.implements.AutoJSONDeserializable or variable.type.implements.JSONDeserializable %}{% else %}{% if variable.type.kind == "enum" and variable.type.rawTypeName %}{{ variable.type.rawTypeName.name }}{% else %}{{ variable.unwrappedTypeName }}{% endif %}{% endif %}{% endset %}
+        {% set itemOperation %}{% if variable.type.implements.AutoJSONDeserializable or variable.type.implements.JSONDeserializable %}.flatMap({{ variable.unwrappedTypeName }}.init(JSONObject:)){%else%}{% if variable.type.kind == "enum" and variable.type.rawTypeName %}.flatMap({ {{ variable.actualTypeName.unwrappedTypeName }}(rawValue: $0) }){% endif %}{% endif %}{% endset %}
         {% set Assignment %}let {{ variable.name }} = (JSONObject["{{ variable.annotations.JSONKey|default:variable.name }}"] {% if castType %}as? {% endif %}{{ castType }}){{ itemOperation }}{% endset %}
         {% if variable.isOptional %}
         {{ Assignment }}
@@ -22,8 +22,8 @@ extension {{ type.name }}: JSONDeserializable {
         {% call Optional Assignment %}
         {% endif %}
         {% else %}
-        {% set castType %}{% if variable.typeName.array.elementType.implements.AutoJSONDeserializable or variable.typeName.array.elementType.implements.JSONDeserializable %}[Any]{% else %}{% if variable.typeName.array.elementTypeName.unwrappedTypeName == "Date" %}[String]{% else %}{% if variable.typeName.array.elementType.kind == "enum" and variable.typeName.array.elementType.rawTypeName %}[{{ variable.typeName.array.elementType.rawTypeName.name }}]{% else %}{{ variable.typeName }}{% endif %}{% endif %}{% endif %}{% endset %}
-        {% set itemOperation %}{% if variable.typeName.array.elementType.implements.AutoJSONDeserializable or variable.typeName.array.elementType.implements.JSONDeserializable %}?.flatMap({{ variable.typeName.array.elementTypeName.unwrappedTypeName }}.init(JSONObject:)){%else%}{% if variable.typeName.array.elementTypeName.unwrappedTypeName == "Date" %}?.flatMap(JSONDateFormatter.date(from:)){% else %}{% if variable.typeName.array.elementType.kind == "enum" and variable.typeName.array.elementType.rawTypeName %}?.flatMap({ {{ variable.typeName.array.elementTypeName.unwrappedTypeName }}(rawValue: $0) }){% endif %}{% endif %}{% endif %}{% endset %}
+        {% set castType %}{% if variable.typeName.array.elementType.implements.AutoJSONDeserializable or variable.typeName.array.elementType.implements.JSONDeserializable %}[Any]{% else %}{% if variable.typeName.array.elementType.kind == "enum" and variable.typeName.array.elementType.rawTypeName %}[{{ variable.typeName.array.elementType.rawTypeName.name }}]{% else %}{{ variable.typeName }}{% endif %}{% endif %}{% endset %}
+        {% set itemOperation %}{% if variable.typeName.array.elementType.implements.AutoJSONDeserializable or variable.typeName.array.elementType.implements.JSONDeserializable %}?.flatMap({{ variable.typeName.array.elementTypeName.unwrappedTypeName }}.init(JSONObject:)){%else%}{% if variable.typeName.array.elementType.kind == "enum" and variable.typeName.array.elementType.rawTypeName %}?.flatMap({ {{ variable.typeName.array.elementTypeName.unwrappedTypeName }}(rawValue: $0) }){% endif %}{% endif %}{% endset %}
         {% set Assignment %}let {{ variable.name }} = (JSONObject["{{ variable.annotations.JSONKey|default:variable.name }}"] as? {{ castType }}){{ itemOperation }}{% endset %}
         {% if variable.isOptional %}
         {{ Assignment }}
